refactor: import timer color names directly
