+++ apps/game-client/src/features/settings/components/timers/components/default-color-item.tsx
@@ -1,11 +1,12 @@
 import { Button } from "@/components/ui/button";
 import { Tile } from "@/components/ui/tile";
 import type { TIMERS_COLORS } from "@/features/timers/constants/timer-colors";
+import { getDefaultColorName } from "@/features/timers/utils/get-default-color-name";
 import { Edit2, Trash2 } from "lucide-react";
 import type { FC } from "react";
 import { useTranslation } from "react-i18next";
 import { ColorEditForm } from "./color-edit-form";
-import { getDefaultColorName, type ColorEditData } from "./color-utils";
+import type { ColorEditData } from "./color-utils";
 
 interface DefaultColorItemProps {
   colorId: string;
